Fix checkmarks being displayed stretched for contact pictures.
Taking the intrinsic size of the contact picture for the whole drawable
has the unwanted side effect of the picture size being applied to the
checkmark drawable as well. Fix it by not applying an intrinsic size to
the flip drawable and doing the scaling formerly done by ImageView by
ourselves.
Also take some cleanup potential.

Change-Id: I0fb07d0631781a345e6a0602ac181a6d76cdf476